Move feature engineering so that it occurs witin the n-fold loop.
Update FeatureFactory to guard if 'id' or 'target' columns don't exist.
Added "Learning Curves (Train vs. Validation AUC)", "Feature Importance (Gain)", and "ROC Curve (Receiver Operating Characteristic)" charts to the XGBoost notebook.  These will be ported to a utility class to be used by the other notebooks.
Note that the LightGBM notebook has an error in it.  This is due to the kernel crashing during the run.  (I couldn't wait for a clean run to check in).